# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Bootcamp - OpenAPI 3.0
  description: Contrato de Banco Bootcamp
  version: 1.0.0
servers:
  - url: nttdata/
...
paths:
...
  /credit:
    get:
      tags:
        - Credit
      summary: Find All Credit
      description: Find All Credit
      operationId: findAllCredit
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponseDto'
    put:
      tags:
        - Credit
      summary: Update an existing Credit
      description: Update an existing Credit by Id
      operationId: updateCredit
      requestBody:
        description: Update an existent Credit in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponseDto'
  /credit/person:
    post:
      tags:
        - Credit
      summary: Add a new Credit person to the bank
      description: Add a new Credit person to the bank
      operationId: createCreditPerson
      requestBody:
        description: Create a new Credit person in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponseDto'
  /credit/company:
    post:
      tags:
        - Credit
      summary: Add a new Credit company to the bank
      description: Add a new Credit company to the bank
      operationId: createCreditCompany
      requestBody:
        description: Create a new Credit company in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponseDto'
  /credit/{creditId}:
    get:
      tags:
        - Credit
      summary: Find Credit by id
      description: Find Credit by id
      operationId: findCreditById
      parameters:
        - name: creditId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditResponseDto'
  /credit/pay:
    post:
      tags:
        - Credit
      summary: Pay a credit person or company to the bank
      description: Pay a credit person or company to the bank
      operationId: payCredit
      requestBody:
        description: Pay a Credit person or company in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditPayRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditPayResponseDto'
  /creditCard:
    get:
      tags:
        - CreditCard
      summary: Find All CreditCard
      description: Find All CreditCard
      operationId: findAllCreditCard
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponseDto'
    put:
      tags:
        - CreditCard
      summary: Update an existing CreditCard
      description: Update an existing CreditCard by Id
      operationId: updateCreditCard
      requestBody:
        description: Update an existent CreditCard in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponseDto'
  /creditCard/person:
    post:
      tags:
        - CreditCard
      summary: Add a new creditCard person to the bank
      description: Add a new creditCard person to the bank
      operationId: createCreditCardPerson
      requestBody:
        description: Create a new CreditCard person in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponseDto'
  /creditCard/company:
    post:
      tags:
        - CreditCard
      summary: Add a new CreditCard company to the bank
      description: Add a new CreditCard company to the bank
      operationId: createCreditCardCompany
      requestBody:
        description: Create a new CreditCard company in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponseDto'
  /creditCard/{creditCardId}:
    get:
      tags:
        - CreditCard
      summary: Find CreditCard by id
      description: Find CreditCard by id
      operationId: findCreditCardById
      parameters:
        - name: creditCardId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponseDto'
  /creditCard/customer/{customerId}:
    get:
      tags:
        - CreditCard
      summary: Find CreditCard by customerId
      description: Find CreditCard by customerId
      operationId: findCreditCardByCustomerId
      parameters:
        - name: customerId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardResponseDto'
  /creditCard/pay:
    post:
      tags:
        - CreditCard
      summary: Pay a creditCard person or company to the bank
      description: Pay a creditCard person or company to the bank
      operationId: payCreditCard
      requestBody:
        description: Pay a CreditCard person or company in the bank
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardPayRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardPayResponseDto'
  /creditCard/consume:
    post:
      tags:
        - CreditCard
      summary: Consume credit card balance
      description: consume credit card balance
      operationId: consumeCreditCard
      requestBody:
        description: consume credit card balance
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditCardConsumeRequestDto'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditCardConsumeResponseDto'
...
components:
  schemas:
...
    MessageStatus:
      required:
        - message
        - statusCode
      type: object
      properties:
        statusCode:
          type: string
          example: "0"
        message:
          type: string
          example: Operacion exitosa.
...
    Credit:
      required:
        - amountFee
        - createDate
        - customerId
        - customerType
        - id
        - payDay
        - quantityFee
        - settlementAmount
        - totalAmount
        - totalFee
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerType:
          type: string
          example: PERSONAL
        totalAmount:
          type: number
          format: double
          example: 500
        amountFee:
          type: number
          format: double
          example: 18
        quantityFee:
          type: integer
          example: 12
        totalFee:
          type: integer
          example: 12
        settlementAmount:
          type: number
          format: double
          example: 500
        payDay:
          type: string
          example: 12-03-2023
        createDate:
          type: string
          example: 12-03-2023
    CreditResponseDto:
      required:
        - credit
        - statusDto
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
        credit:
          type: array
          items:
            $ref: '#/components/schemas/Credit'
    CreditRequestDto:
      required:
        - amountFee
        - createDate
        - customerId
        - customerType
        - payDay
        - quantityFee
        - settlementAmount
        - totalAmount
        - totalFee
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerType:
          type: string
          example: PERSONAL
        totalAmount:
          type: number
          format: double
          example: 500
        amountFee:
          type: number
          format: double
          example: 18
        quantityFee:
          type: integer
          example: 12
        totalFee:
          type: integer
          example: 12
        settlementAmount:
          type: number
          format: double
          example: 500
        payDay:
          type: string
          example: 12-03-2023
        createDate:
          type: string
          example: 12-03-2023
    CreditPayRequestDto:
      required:
        - amount
        - customerId
        - id
        - payDay
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        amount:
          type: number
          format: double
          example: 500
        payDay:
          type: string
          example: 12-03-2023
    CreditPayResponseDto:
      required:
        - statusDto
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
    CreditCard:
      required:
        - createDate
        - creditCardAccount
        - creditCardId
        - creditLine
        - customerId
        - feeMonth
        - id
        - lineAvailable
        - minimumAmount
        - payDay
        - settlementAmount
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        creditLine:
          type: number
          format: double
          example: 1000
        creditCardAccount:
          type: string
          example: 6413619bf92d28ec5a2e104e
        creditCardId:
          type: string
          example: 1111-2222-3333-4444
        lineAvailable:
          type: number
          format: double
          example: 1000
        feeMonth:
          type: number
          format: double
          example: 0
        minimumAmount:
          type: number
          format: double
          example: 0
        settlementAmount:
          type: number
          format: double
          example: 0
        payDay:
          type: string
          example: 12-03-2023
        createDate:
          type: string
          example: 12-03-2023
    CreditCardResponseDto:
      required:
        - creditCard
        - statusDto
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
        creditCard:
          type: array
          items:
            $ref: '#/components/schemas/CreditCard'
    CreditCardRequestDto:
      required:
        - createDate
        - creditCardAccount
        - creditCardId
        - creditLine
        - customerId
        - feeMonth
        - lineAvailable
        - minimumAmount
        - payDay
        - settlementAmount
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        creditLine:
          type: number
          format: double
          example: 1000
        creditCardAccount:
          type: string
          example: 6413619bf92d28ec5a2e104e
        creditCardId:
          type: string
          example: 1111-2222-3333-4444
        lineAvailable:
          type: number
          format: double
          example: 1000
        feeMonth:
          type: number
          format: double
          example: 0
        minimumAmount:
          type: number
          format: double
          example: 0
        settlementAmount:
          type: number
          format: double
          example: 0
        payDay:
          type: string
          example: 12-03-2023
        createDate:
          type: string
          example: 12-03-2023
    CreditCardPayRequestDto:
      required:
        - amount
        - customerId
        - id
        - payDay
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        amount:
          type: number
          format: double
          example: 500
        payDay:
          type: string
          example: 12-03-2023
    CreditCardPayResponseDto:
      required:
        - statusDto
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
    CreditCardConsumeRequestDto:
      required:
        - amount
        - customerId
        - id
      type: object
      properties:
        id:
          type: string
          example: 6413619bf92d28ec5a2e104e
        customerId:
          type: string
          example: 6413619bf92d28ec5a2e104e
        amount:
          type: number
          format: double
          example: 500
    CreditCardConsumeResponseDto:
      required:
        - statusDto
      type: object
      properties:
        statusDto:
          $ref: '#/components/schemas/MessageStatus'
...